Made zina and dialogs into sazina.
