+++ ServiceLayer/Interface/ISazinaManager.cs
@@ -3,10 +3,11 @@ using MentalaisGidsAPI.Domain;
 
 namespace ServiceLayer.Interface
 {
-    public interface IZinaManager : IBaseManager<Zina>
+    public interface ISazinaManager : IBaseManager<Dialogs>
     {
+        Task<bool> StopDialogue(int dialogueId);
+        Task<bool> StartDialogue(int receiverId);
         Task<List<ZinaDto>> GetZinas(int dialogsId);
         Task<bool> PostZina(int receiverId, string zina);
     }
 }
-
