+++ ServiceLayer/Manager/SazinaManager.cs
@@ -11,12 +11,12 @@ using System.Threading.Tasks;
 
 namespace ServiceLayer.Manager
 {
-    public class DialogsManager : BaseManager<Dialogs>, IDialogsManager
+    public class SazinaManager : BaseManager<Dialogs>, ISazinaManager
     {
         private readonly MentalaisGidsContext _context;
         private readonly IUserService _userService;
 
-        public DialogsManager(MentalaisGidsContext context, IUserService userService) : base(context)
+        public SazinaManager(MentalaisGidsContext context, IUserService userService) : base(context)
         {
             _context = context;
             _userService = userService;
@@ -74,6 +74,53 @@ namespace ServiceLayer.Manager
             };
 
             await SaveOrUpdate(newDialogue);
+            return true;
+        }
+        public async Task<List<ZinaDto>?> GetZinas(int dialogsId)
+        {
+            var zinas = await _context.Zina
+                          .Include(x => x.Autors)
+                          .Where(x => x.DialogsID == dialogsId)
+                          .ToListAsync();
+
+            if (zinas == null)
+            {
+                return null;
+            }
+
+            var dtoList = zinas.Select(zina => new ZinaDto
+            {
+                DatumsUnLaiks = zina.DatumsUnLaiks,
+                Saturs = zina.Saturs,
+                AutoraVards = zina.Autors.Vards,
+                AutoraUzvards = zina.Autors.Uzvards
+            }).ToList();
+
+            return dtoList;
+        }
+        public async Task<bool> PostZina(int dialogsId, string zina)
+        {
+            var senderId = _userService.GetUserId();
+            var dialogue = await _context.Dialogs.FirstOrDefaultAsync(x => x.DialogsID == dialogsId);
+
+            if (dialogue == null)
+            {
+                return false;
+            }
+
+
+            var zinaObj = new Zina
+            {
+                DialogsID = dialogue.DialogsID,
+                AutorsID = senderId,
+                DatumsUnLaiks = DateTime.Now,
+                Saturs = zina
+            };
+
+
+            await _context.Zina.AddAsync(zinaObj);
+            await _context.SaveChangesAsync();
+
             return true;
         }
     }
