+++ ServiceLayer/SajutuNovertejumsManager.cs
@@ -4,6 +4,7 @@ using MentalaisGidsAPI.Domain.dto;
 using MentalaisGidsAPI.Domain.Dto;
 using Microsoft.EntityFrameworkCore;
 using ServiceLayer.Interface;
+using ServiceLayer.Manager;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory;
